Update TFile.pde
Added TThumbnailable in the hierarchy of TFile.

TThumbnailable is a child of TFile.
TMovie and TImage both extend it, and thus both automatically have the getThumbnailAbsolutePath method - no need to implement it twice! (avoiding code duplication was the motivation for this change / commit)
I also wanted to avoid putting getThumbnailAbsoluteLocation method directly in TFile, since that wouldn't make sense for, say, a NoFile. Thus, a new TFile was made; the TThumbnailable!