Fixed Redraw Issues
The issue was that parts of the frame did not show what was meant to be there. Noticable during scorlling. Parts of the frame still showed images from were the previous scroll state was.

The issue was more noticable at low FPS (for instance in fullscreen())

This occured because all events (mouse and keyboard) are actually a part of a frames drawing process, so calling redraw from within mouseWheel for instance would not actually make processing draw a new frame again.

The added reedraw function enusres this is actually done. A (minor) disadvantage is that now two draw calls are really done for every redraw (although only the first one does any graphics stuffs, the second is there only to ensure that nothing new has happened while the first draw was taking place, and if this indeed the case, the second draw will end early)

Thtat's a wrap! Get u some hot fixes.