Resize Event Detection
T'was a hastle, but it's done.
Processing does not (yet, see https://github.com/processing/processing/tree/master/core and look for resize event) provide a way to detect a change in the window size. This probably for good reason, so I doubt this hack is very stable. Oh well, it works for me, and this is the work I've done for now

Currently, resizing works when an file is being viewed in fullscreen.

** How will the contextHistory react to a window size change??